Added saveClass and saveAssignment activities. Refactored saveCategory activity. Refactored DatabaseContract. Commented out button in the main activity.
