+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.AddCategoryActivity"></activity>
+        <activity android:name=".activities.SaveCategoryActivity" />
+        <activity android:name=".activities.SaveClassActivity" />
+        <activity android:name=".activities.SaveAssignmentActivity"></activity>
     </application>
 
 </manifest>
