+++ app/src/main/java/com/projectdolphin/activities/SaveAssignmentActivity.java
@@ -1,0 +1,15 @@
+package com.projectdolphin.activities;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.projectdolphin.R;
+
+public class SaveAssignmentActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_save_assignment);
+    }
+}
