+++ app/src/main/java/com/projectdolphin/activities/SaveClassActivity.java
@@ -1,0 +1,15 @@
+package com.projectdolphin.activities;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.projectdolphin.R;
+
+public class SaveClassActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_add_class);
+    }
+}
