+++ app/src/main/res/layout/view_card.xml
@@ -45,13 +45,14 @@
                     android:layout_alignParentRight="true"
                     android:layout_alignParentEnd="true"/>
 
-                <Button
-                    android:text="ADD CATEGORY"
-                    android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:layout_alignParentRight="true"
-                    android:layout_alignParentEnd="true"
-                    android:onClick="onAddCategory"/>
+                <!-- This button starts my class for now -->
+                <!--<Button-->
+                    <!--android:text="ADD CATEGORY"-->
+                    <!--android:layout_width="wrap_content"-->
+                    <!--android:layout_height="match_parent"-->
+                    <!--android:layout_alignParentRight="true"-->
+                    <!--android:layout_alignParentEnd="true"-->
+                    <!--android:onClick="onAddCategory"/>-->
 
             </RelativeLayout>
             <RelativeLayout
